+++ srcs/pixel_point/pixels_1.c
@@ -37,46 +37,7 @@ int		rgba(int r, int g, int b, int a)
 	return ((a << 24) | (r << 16) | (g << 8) | (b));
 }
 
-int		argb(int color)
-{
-	int new;
-
-	new = ((color & 0xff) << 24) \
-	| (((color >> 8) & 0xff) << 16) \
-	| (((color >> 16) & 0xff) << 8) \
-	| ((color >> 24) & 0xff);
-	return (new);
-}
-
-int		abgr_inversion(int r, int g, int b, int a)
-{
-	return ((a << 24) | (b << 16) | (g << 8) | (r));
-}
-
-int		pack_color_channels(int fst, int snd, int trd, int fth)
-{
-	return ((fst << 24) | (snd << 16) | (trd << 8) | (fth));
-}
-
-int	avg_colour(int start, int end, int num, int den)
-{
-	t_ulong	r;
-	t_ulong	g;
-	t_ulong	b;
-	t_ulong	a;
-
-	if (den == 0)
-		return (start);
-	r = rgb_r(start) * num / den \
-		+ rgb_r(end) * (den - num) / den;
-	g = rgb_g(start) * num / den \
-		+ rgb_g(end) * (den - num) / den;
-	b = rgb_b(start) * num / den \
-		+ rgb_b(end) * (den - num) / den;
-	a = rgb_a(start);
-	return (pack_color_channels((int)r, (int)g, (int)b, (int)a));
-}
-
+/*
 int	avg_colour_new(int start, int end, int num, int den)
 {
 	int		new;
@@ -104,7 +65,7 @@ int	avg_colour_3(int start, int end, int num, int den)
 	char	*ptr_new;
 	char	pix_start;
 	char	pix_end;
-	
+
     if (den == 0)
 		return (start);
     ptr_new = (char *)&new;
@@ -118,3 +79,4 @@ int	avg_colour_3(int start, int end, int num, int den)
 	}
 	return (*(int *)ptr_new);
 }
+*/
