+++ srcs/pixel_point/translate_rotate.c
@@ -12,7 +12,6 @@
 
 #include "compass.h"
 
-
 void	translate_point(t_pixel	*point, int dx, int dy)
 {
 	point->x += dx;
@@ -22,22 +21,22 @@ void	translate_point(t_pixel	*point, int dx, int dy)
 void	backward_rotate_point(t_pixel *point, t_pixel centre, \
 	float cos, float sin)
 {
-	t_pixel og;
+	t_pixel	og;
 
 	og = *point;
-    point->x = (int)roundf(cos * (og.x - centre.x) + sin \
+	point->x = (int)roundf(cos * (og.x - centre.x) + sin \
 		* (og.y - centre.y) + centre.x);
-    point->y = (int)roundf(sin * (og.x - centre.x) - cos \
+	point->y = (int)roundf(sin * (og.x - centre.x) - cos \
 		* (og.y - centre.y) + centre.y);
 }
 
 void	rotate_point(t_pixel *point, t_pixel centre, float cos, float sin)
 {
-	t_pixel og;
+	t_pixel	og;
 
 	og = *point;
-    point->x = (int)(cos * (og.x - centre.x) \
+	point->x = (int)(cos * (og.x - centre.x) \
 		+ sin * (og.y - centre.y) + centre.x);
-    point->y = (int)(- sin * (og.x - centre.x) \
+	point->y = (int)(-sin * (og.x - centre.x) \
 		+ cos * (og.y - centre.y) + centre.y);
 }
